Orders/requirements.txt
asgiref==3.4.1
distlib==0.3.6
Django==3.2.19
djangorestframework==3.14.0
filelock==3.4.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
platformdirs==2.4.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.1.1
virtualenv==20.17.1
zipp==3.6.0
